Fully support WITH_CYCLES_LOGGING option
Quite straightforward, used find*() modules from Ceres, so the
variable naming is a bit different. Example of declaring which
Glog/Gflags to use:

  cmake \
    -D GLOG_INCLUDE_DIR_HINTS=/opt/lib/glog/include \
    -D GLOG_LIBRARY_DIR_HINTS=/opt/lib/glog/lib \
    -D GFLAGS_INCLUDE_DIR_HINTS=/opt/lib/gflags/include \
    -D GFLAGS_LIBRARY_DIR_HINTS=/opt/lib/gflags/lib \
    /path/to/cycles

This would break logging support when re-bundled back to Blender
because of the changes in CMake files. But those changes i would
like to do in Blender as well with some tweaks to how Glog/Gflags
libraries are being detected.