Add tangent generation in the Cycles core
This makes it available in Cycles standalone, and the implementation
can be shared with Blender.

Additionally this adds support for OSL parameter metadata named
`defaultgeomprop`, whose values are interpreted the same way as
the property on MaterialX node inputs. When set to `Tworld` the
tangent is then automatically linked to the shader and generated for
the mesh.

Pull Request: https://projects.blender.org/blender/cycles/pulls/25