Fix: Cycles: NaN Normals from Normal Map
a degenerate triangle could produce a tangent that is antiparallel to
the normal, resulting the mapped normal to be zero, and becomes NaN when
normalized in `object_normal_transform()`. Fixed by falling back to
unperturbed normal in this case.

Fixes an assertion in the attic benchmark scene.

Pull Request: https://projects.blender.org/blender/blender/pulls/140135